Hide the DESTDIR environment variable from cmake, otherwise the build fails
I have DESTDIR set so that I can influence where my own application is
installed when running "make install", but this affects the cmake build of
turbojpeg-sys by installing files in the wrong place.

$ mkdir /dev/shm/DESTDIR
$ DESTDIR=/dev/shm/DESTDIR cargo build
   Compiling turbojpeg-sys v1.0.1 (/home/simon/src/rust-turbojpeg/turbojpeg-sys)
error: could not find native static library `turbojpeg`, perhaps an -L flag is missing?

error: could not compile `turbojpeg-sys` (lib) due to 1 previous error

$ find /dev/shm/DESTDIR/
/dev/shm/DESTDIR/
/dev/shm/DESTDIR/home
/dev/shm/DESTDIR/home/simon
/dev/shm/DESTDIR/home/simon/src
/dev/shm/DESTDIR/home/simon/src/rust-turbojpeg
/dev/shm/DESTDIR/home/simon/src/rust-turbojpeg/turbojpeg-sys
/dev/shm/DESTDIR/home/simon/src/rust-turbojpeg/turbojpeg-sys/target
/dev/shm/DESTDIR/home/simon/src/rust-turbojpeg/turbojpeg-sys/target/debug
/dev/shm/DESTDIR/home/simon/src/rust-turbojpeg/turbojpeg-sys/target/debug/build
/dev/shm/DESTDIR/home/simon/src/rust-turbojpeg/turbojpeg-sys/target/debug/build/turbojpeg-sys-3cb1ee0bda3dbcb0

Overriding DESTDIR to be "" fixes the problem. This could be considered a
bug in the cmake-rs API? At the very least it needs to be possible to
configure it to use a clean environment to ensure consistent results.